+++ src/mask/none.rs
@@ -1,4 +1,5 @@
 use crate::mask::Mask;
+
 /// A mask where every element is set.
 #[derive(Default, Debug, Clone, Copy)]
 pub struct None(());
@@ -6,7 +7,7 @@ pub struct None(());
 impl Mask for None {
     type Iter = Iter;
 
-    fn test(&self, _: usize) -> bool {
+    fn test(&self, _: u32) -> bool {
         false
     }
 
@@ -19,7 +20,7 @@ impl Mask for None {
 pub struct Iter(());
 
 impl Iterator for Iter {
-    type Item = usize;
+    type Item = u32;
 
     fn next(&mut self) -> Option<Self::Item> {
         Option::None
